# internal/server/web/openapi.yaml
...
openapi: 3.1.0
info:
  title: wudict
  version: "1"
  summary: A local dictionary server for MDX/MDD, StarDict, Slob, DSL and BGL files.
  description: |
    WuWeiDict serves your own dictionary files over HTTP, on your own machine.

    **Streaming.** `/api/dicts`, `/api/search` and `/api/rescan` answer with
    NDJSON: one JSON object per line, flushed the moment it is ready. Read them
    line by line; do not wait for the body to end. That is the whole point - the
    first dictionary's results arrive while the slowest one is still opening.
    Every stream opens with a `begin` line and closes with an `end` line.

    `/api/ingest` streams Server-Sent Events instead, because it reports
    progress rather than results.

    **Switches.** Every boolean query parameter reads the same spellings: off
    for `0`, `false`, `no`, `off` or an empty value, on for anything else.

    **Two surfaces.** Three read-only endpoints answer a browser extension
    cross-origin: `/api/dicts`, `/api/search` and `/res/`. Everything else is
    same-origin, is marked `internal` below, and may change between releases.

    **An access key, when the server is reachable.** On the default
    `127.0.0.1` bind there is nothing to authenticate and no key is asked for.
    On any other address - and Go's wildcard listen is dual-stack, so `0.0.0.0`
    includes every globally routable IPv6 address the machine holds - `AUTH`
...
    and let the browser keep the cookie it sets. Without it: `401`, with
...

...

...
servers:
...
security: []
...
paths:
...
  /api/clear-database:
    post:
      tags: [internal]
      operationId: clearDatabase
      x-internal: true
      summary: Clear selected dictionary indexes and obsolete prepared data
      description: >
        Requires local access or ALLOW_REMOTE_DELETE. Source dictionaries and
        their resources are preserved. Index rebuilds atomically replace the
        main index, retaining optional indexes unless selected for removal.
        Obsolete cleanup includes unused prepared folders regardless of keep
        markers, duplicates, incomplete indexes, abandoned ingest files and
        packed media. Obsolete active indexes are refreshed from their sources.
        Database dictionaries supplied as sources are never cleared.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                index: { type: boolean, default: false }
                contains: { type: boolean, default: false }
                fullText: { type: boolean, default: false }
                obsolete: { type: boolean, default: false }
      responses:
        '200':
          description: Completed; failures are reported per item.
          content:
            application/json:
              schema:
                type: object
                required: [failed]
                properties:
                  failed:
                    type: array
                    items: { type: string }
        '400':
          description: Invalid request or no option selected.
        '403':
          description: Removal permission denied.
  /api/rescan:
    post:
      tags: [internal]
      operationId: updateDictionaryIndexes
      x-internal: true
      summary: Rescan folders and apply the selected index actions
      description: >
        New dictionaries receive a main index and only the selected optional
        indexes, respecting configured DSL variants. Existing optional indexes
        are updated only when already present; deletion cancels pending creation.
        Keep preserves existing feature presence. Cleanup always removes unused
        or disabled prepared indexes and legacy packed media. Sources survive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new, existing]
              properties:
                new:
                  type: object
                  required: [index]
                  properties:
                    index: { type: boolean, enum: [true] }
                    contains: { type: boolean }
                    fullText: { type: boolean }
                existing:
                  type: object
                  properties:
                    index: { type: string, enum: [recreate, keep] }
                    contains: { type: string, enum: [update, delete, keep] }
                    fullText: { type: string, enum: [update, delete, keep] }
      responses:
        '200':
          description: Completed; failed lists any individual failures.
          content:
            application/json:
              schema:
                type: object
                properties:
                  failed: { type: array, items: { type: string } }
        '400':
          description: Invalid index actions.
        '403':
          description: Removal permission denied.
    get:
      tags: [internal]
      operationId: rescan
      x-internal: true
      summary: Re-scan the dictionary folders, then list
      description: >
        Reads the dictionary folders again, warms the registry, and then answers
        exactly as /api/dicts does. Removes unused prepared folders (including
        previously kept caches), abandoned indexes and packed media. Source
        files and both discovered DSL variants are preserved. Requires removal
        permission, like POST /api/clear-database.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            # picks up files added to the dictionary folders, then lists
            curl -sN 'http://127.0.0.1:6888/api/rescan' | head -1
      responses:
        '200':
          description: The /api/dicts stream, after the rescan.
          content:
            application/x-ndjson:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DictsBegin'
                  - $ref: '#/components/schemas/DictsRow'
                  - $ref: '#/components/schemas/StreamEnd'
                discriminator: { propertyName: t }
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
...
    ServerError:
      description: The dictionary could not be opened, or the stream could not start.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Every failing request answers with this shape.
      properties:
        error: { type: string }
      required: [error]
...
    Caps:
      type: object
      description: >
        Which search modes this dictionary can serve now. The authority on what
        to offer: Contains and FTS stay false until that index is built. Field
        names are capitalised on the wire.
      properties:
        Exact: { type: boolean }
        Prefix: { type: boolean }
        Contains: { type: boolean }
        FTS: { type: boolean }
...
    StreamEnd:
      type: object
      description: The last line of any NDJSON stream.
      properties:
        t: { const: end }
        i:
          type: integer
          description: Always 0 here. The search stream carries the field on every line.
      required: [t]
...
    DictsBegin:
      type: object
      properties:
        t: { const: begin }
        total: { type: integer, description: How many rows follow. }
        groupsProblems:
          type: integer
          description: >
            How many lines or items groups.ini skipped (`GET /api/groups` lists
            them); absent when none.
      required: [t]
    DictsRow:
      type: object
      properties:
        t: { const: dict }
        dict: { $ref: '#/components/schemas/DictInfo' }
      required: [t, dict]
    DictInfo:
      type: object
      description: One dictionary, as the list and the panel see it.
      properties:
        job:
          type: object
          description: >
            A change to this dictionary's indexes running now (`/api/ingest`
            follows it); absent when none runs.
          properties:
            done: { type: integer }
            total: { type: integer }
        id: { type: string, description: The value to pass as `dict` to /api/search. }
        name: { type: string }
        format:
          type: string
          enum: [mdx, stardict, slob, dsl, dsl-gd, bgl, zim, wmd, wudict]
        unavailable:
          type: boolean
          description: Disabled DSL variant, retained in settings but not searched.
        dsl:
          type: object
          description: Shared DSL family and effective variant selection.
          properties:
            source: { type: string }
            variant: { type: string, enum: [original, gd] }
            mode: { type: string, enum: [original, gd, both] }
            original: { type: boolean, description: Original variant exists. }
            gd: { type: boolean, description: GD variant exists. }
            sourceAvailable: { type: boolean, description: Shared DSL source exists and can rebuild deleted indexes. }
            indexRemoved: { type: boolean, description: Explicitly removed index stays unavailable until explicit ingest; persists across restart. }
        path:
          type: string
          description: The source file, or the text.db of a prepared dictionary.
        entries: { type: integer, description: Headword count. }
        caps: { $ref: '#/components/schemas/Caps' }
        dbPath: { type: string, description: The prepared database. }
        error: { type: string, description: This dictionary could not be opened. }
        containsStale:
          type: boolean
          description: >
            The substring index predates the current text folding, so it may
            miss words. It keeps working; the panel offers a rebuild.
        outdated:
          type: boolean
          description: >
            The prepared data no longer matches what this build would prepare
            from the source: an older build wrote it, the source changed, or it
            cannot be read. Never set when the source is gone. Reported only;
            `POST /api/reindex` rebuilds every such dictionary.
        source: { type: string, description: 'The original file, if still on disk.' }
        mediaSrc:
          type: array
          items: { type: string }
          description: Companion media sources, such as .mdd files.
        textDB: { type: string }
        mediaDB: { type: string }
        folder:
          type: string
          description: The library folder holding them - the transferable unit.
        dbSize: { type: integer, description: Bytes of text.db. }
        mediaSize: { type: integer, description: Bytes of media.db. }
        hasMedia: { type: boolean, description: 'Packable media exists, so "pack media" is offered.' }
        builtin: { type: boolean, description: 'Shipped with the app (the wudict howto, id `wudict-howto`), not added by the user; it cannot be removed.' }
        groups:
          type: array
          items: { $ref: '#/components/schemas/DictGroup' }
          description: >
            The groups this dictionary belongs to, which the picker offers as
            single choices instead of a hundred names. Derived per request from
            this row's own name, path and declared language - never stored -
            and absent when nothing could be established, which is a normal
            answer. A client selects a group by intersecting the rows itself;
            there is no server-side group query.
        articleLang:
          type: string
          description: >
            ISO 639-1 code of the language the articles are written in - the
            default voice for reading a selection aloud. Derived like groups:
            the declared contents language, else the second code of a language
            pair in the title, file name or an enclosing folder, else (no pair
            anywhere) the headword language. Absent when nothing says.
      required: [id, name, format, path, entries, caps]
    DictGroup:
      type: object
      description: One (facet, value) membership, with the labels to show for it.
      properties:
        f: { type: string, description: 'Facet id: lang | pair | kind | pub.' }
        fl: { type: string, description: 'Facet label, heading the group in the picker.' }
        fo: { type: integer, description: 'Facet rank, ascending; may be negative. The user''s own groups.ini sections rank above Language, the built-in ones after Language pair.' }
        v: { type: string, description: 'Value id, unique within the facet. For pair: two ISO 639-1 codes in code order, the same for either direction ("en-ru"), or one code for a stated monolingual pair ("es").' }
        vl: { type: string, description: 'Value label, naming the choice.' }
      required: [f, fl, fo, v, vl]
...